+++ test/storage.spec.js
@@ -5,7 +5,7 @@ const chai = require('chai');
 const expect = chai.expect;
 const GridFsStorage = require('../index');
 const setting = require('./utils/settings');
-const { files, cleanStorage, getDb, getClient } = require('./utils/testutils');
+const {files, cleanStorage, getDb, getClient} = require('./utils/testutils');
 const request = require('supertest');
 const multer = require('multer');
 const mongo = require('mongodb');
@@ -13,22 +13,25 @@ const MongoClient = mongo.MongoClient;
 const md5File = require('md5-file');
 const fs = require('fs');
 const Promise = require('bluebird');
+const cache = GridFsStorage.cache;
 
 chai.use(require('chai-interface'));
 
-describe('Storage', function () {
+describe('Storage', () => {
   let result, app, storage;
 
   before(() => app = express());
 
-  describe('url created instance', function () {
-    before((done) => {
-      storage = new GridFsStorage({ url: setting.mongoUrl() });
+  beforeEach(() => cache.clear());
 
-      const upload = multer({ storage });
+  describe('url created instance', () => {
+    before((done) => {
+      storage = new GridFsStorage({url: setting.mongoUrl});
+
+      const upload = multer({storage});
 
       app.post('/url', upload.array('photos', 2), (req, res) => {
-        result = { headers: req.headers, files: req.files, body: req.body };
+        result = {headers: req.headers, files: req.files, body: req.body};
         res.end();
       });
 
@@ -41,7 +44,7 @@ describe('Storage', function () {
       });
     });
 
-    it('should store the files on upload', function () {
+    it('should store the files on upload', () => {
       expect(result.files).to.be.an('array');
       expect(result.files).to.have.length(2);
     });
@@ -57,18 +60,18 @@ describe('Storage', function () {
 
   });
 
-  describe('db created instance', function () {
+  describe('db created instance', () => {
     before(() => {
-      const promise = MongoClient.connect(setting.mongoUrl());
+      const promise = MongoClient.connect(setting.mongoUrl);
       return promise.then(_db => {
         const db = getDb(_db);
-        storage = GridFsStorage({ db });
+        storage = GridFsStorage({db});
         storage.client = getClient(_db);
 
-        const upload = multer({ storage });
+        const upload = multer({storage});
 
         app.post('/db', upload.array('photos', 2), (req, res) => {
-          result = { headers: req.headers, files: req.files, body: req.body };
+          result = {headers: req.headers, files: req.files, body: req.body};
           res.end();
         });
 
@@ -82,7 +85,7 @@ describe('Storage', function () {
       });
     });
 
-    it('should store the files on upload', function () {
+    it('should store the files on upload', () => {
       expect(result.files).to.be.an('array');
       expect(result.files).to.have.length(2);
     });
@@ -98,23 +101,23 @@ describe('Storage', function () {
 
   });
 
-  describe('db promise based instance', function () {
+  describe('db promise based instance', () => {
     let db, client;
     before((done) => {
 
       const promised = MongoClient
-        .connect(setting.mongoUrl()).then(_db => {
+        .connect(setting.mongoUrl).then(_db => {
           db = getDb(_db);
           client = getClient(_db);
           return db;
         });
 
-      storage = GridFsStorage({ db: promised });
-      const upload = multer({ storage });
+      storage = GridFsStorage({db: promised});
+      const upload = multer({storage});
 
 
       app.post('/promise', upload.array('photos', 2), (req, res) => {
-        result = { headers: req.headers, files: req.files, body: req.body };
+        result = {headers: req.headers, files: req.files, body: req.body};
         res.end();
       });
 
@@ -127,12 +130,12 @@ describe('Storage', function () {
       });
     });
 
-    it('should store the files on upload', function () {
+    it('should store the files on upload', () => {
       expect(result.files).to.be.an('array');
       expect(result.files).to.have.length(2);
     });
 
-    it('should have each stored file the same MD5 signature than the uploaded file', function () {
+    it('should have each stored file the same MD5 signature than the uploaded file', () => {
       result.files.forEach((file, index) => {
         expect(file.md5).to.be.equal(md5File(files[index]));
       });
@@ -141,11 +144,11 @@ describe('Storage', function () {
     after(() => cleanStorage(storage, db, client));
   });
 
-  describe('handle incoming files while connecting', function () {
+  describe('handle incoming files while connecting', () => {
     let db, client;
     before((done) => {
       const promised = MongoClient
-        .connect(setting.mongoUrl()).then((_db) => {
+        .connect(setting.mongoUrl).then((_db) => {
           db = getDb(_db);
           client = getClient(_db);
           return new Promise((resolve) => {
@@ -153,12 +156,12 @@ describe('Storage', function () {
           });
         });
 
-      storage = GridFsStorage({ db: promised });
-      const upload = multer({ storage });
+      storage = GridFsStorage({db: promised});
+      const upload = multer({storage});
 
 
       app.post('/incoming', upload.array('photos', 2), (req, res) => {
-        result = { headers: req.headers, files: req.files, body: req.body };
+        result = {headers: req.headers, files: req.files, body: req.body};
         res.end();
       });
 
@@ -169,12 +172,12 @@ describe('Storage', function () {
         .end(done);
     });
 
-    it('should store the files on upload', function () {
+    it('should store the files on upload', () => {
       expect(result.files).to.be.an('array');
       expect(result.files).to.have.length(2);
     });
 
-    it('should have each stored file the same MD5 signature than the uploaded file', function () {
+    it('should have each stored file the same MD5 signature than the uploaded file', () => {
       result.files.forEach((file, index) => {
         expect(file.md5).to.be.equal(md5File(files[index]));
       });
@@ -184,14 +187,14 @@ describe('Storage', function () {
 
   });
 
-  describe('default uploaded file spec', function () {
+  describe('default uploaded file spec', () => {
     let size;
     before((done) => {
-      storage = GridFsStorage({ url: setting.mongoUrl() });
-      const upload = multer({ storage });
+      storage = GridFsStorage({url: setting.mongoUrl});
+      const upload = multer({storage});
 
       app.post('/spec', upload.single('photo'), (req, res) => {
-        result = { headers: req.headers, file: req.file, body: req.body };
+        result = {headers: req.headers, file: req.file, body: req.body};
         res.end();
       });
 
@@ -208,33 +211,33 @@ describe('Storage', function () {
       });
     });
 
-    it('should have a filename property', function () {
+    it('should have a filename property', () => {
       expect(result.file).to.have.a.property('filename');
       expect(result.file.filename).to.be.a('string');
       expect(result.file.filename).to.match(/^[0-9a-f]{32}$/);
     });
 
-    it('should have a metadata property', function () {
+    it('should have a metadata property', () => {
       expect(result.file).to.have.a.property('metadata');
       expect(result.file.metadata).to.equal(null);
     });
 
-    it('should have a id property', function () {
+    it('should have a id property', () => {
       expect(result.file).to.have.a.property('id');
       expect(result.file.id).to.match(/^[0-9a-f]{24}$/);
     });
 
-    it('should have a size property with the length of the file', function () {
+    it('should have a size property with the length of the file', () => {
       expect(result.file).to.have.a.property('size');
       expect(result.file.size).to.equal(size);
     });
 
-    it('should have the default bucket name pointing to the fs collection', function () {
+    it('should have the default bucket name pointing to the fs collection', () => {
       expect(result.file).to.have.a.property('bucketName');
       expect(result.file.bucketName).to.equal('fs');
     });
 
-    it('should have the date of the upload', function () {
+    it('should have the date of the upload', () => {
       expect(result.file).to.have.a.property('uploadDate');
     });
 
@@ -242,7 +245,7 @@ describe('Storage', function () {
 
   });
 
-  describe('file function usage', function () {
+  describe('file function usage', () => {
     before((done) => {
       let counter = 0;
       const data = ['foo', 'bar'];
@@ -250,8 +253,8 @@ describe('Storage', function () {
       const names = ['plants', 'animals'];
       const contentTypes = ['text/plain', 'image/jpeg'];
       storage = new GridFsStorage({
-        url: setting.mongoUrl(),
-        file: function () {
+        url: setting.mongoUrl,
+        file: () => {
           counter++;
           return {
             filename: 'file' + counter,
@@ -259,15 +262,16 @@ describe('Storage', function () {
             id: counter,
             chunkSize: sizes[counter - 1],
             bucketName: names[counter - 1],
-            contentType: contentTypes[counter - 1]
+            contentType: contentTypes[counter - 1],
           };
-        }
+        },
+
       });
 
-      const upload = multer({ storage });
+      const upload = multer({storage});
 
       app.post('/config', upload.array('photos', 2), (req, res) => {
-        result = { headers: req.headers, files: req.files, body: req.body };
+        result = {headers: req.headers, files: req.files, body: req.body};
         res.end();
       });
 
@@ -280,43 +284,43 @@ describe('Storage', function () {
       });
     });
 
-    it('should the request contain the two uploaded files', function () {
+    it('should the request contain the two uploaded files', () => {
       expect(result.files).to.be.an('array');
       expect(result.files).to.have.length(2);
     });
 
-    it('should be named with the provided value', function () {
+    it('should be named with the provided value', () => {
       expect(result.files[0].filename).to.equal('file1');
       expect(result.files[1].filename).to.equal('file2');
     });
 
-    it('should contain a metadata object with the provided object', function () {
+    it('should contain a metadata object with the provided object', () => {
       expect(result.files[0].metadata).to.equal('foo');
       expect(result.files[1].metadata).to.equal('bar');
     });
 
-    it('should be stored with the provided chunkSize value', function () {
+    it('should be stored with the provided chunkSize value', () => {
       expect(result.files[0].chunkSize).to.equal(102400);
       expect(result.files[1].chunkSize).to.equal(204800);
     });
 
-    it('should change the id with the provided value', function () {
+    it('should change the id with the provided value', () => {
       expect(result.files[0].id).to.equal(1);
       expect(result.files[1].id).to.equal(2);
     });
 
     it('should be stored under in a collection with the provided value', function (done) {
       const db = storage.db;
-      db.collection('plants.files', { strict: true }, function (err) {
+      db.collection('plants.files', {strict: true}, function (err) {
         expect(err).to.be.equal(null);
-        db.collection('animals.files', { strict: true }, function (err) {
+        db.collection('animals.files', {strict: true}, function (err) {
           expect(err).to.be.equal(null);
           done();
         });
       });
     });
 
-    it('should change the content type with the provided value', function () {
+    it('should change the content type with the provided value', () => {
       expect(result.files[0].contentType).to.equal('text/plain');
       expect(result.files[1].contentType).to.equal('image/jpeg');
     });
@@ -324,21 +328,21 @@ describe('Storage', function () {
     after(() => cleanStorage(storage));
   });
 
-  describe('Missing properties in file naming function', function () {
+  describe('Missing properties in file naming function', () => {
     before((done) => {
       storage = GridFsStorage({
-        url: setting.mongoUrl(),
-        file: function () {
+        url: setting.mongoUrl,
+        file: () => {
           return {
-            metadata: { foo: 'bar' },
-            id: 1234
+            metadata: {foo: 'bar'},
+            id: 1234,
           };
-        }
+        },
       });
-      const upload = multer({ storage });
+      const upload = multer({storage});
 
       app.post('/missing', upload.single('photo'), (req, res) => {
-        result = { headers: req.headers, file: req.file, body: req.body };
+        result = {headers: req.headers, file: req.file, body: req.body};
         res.end();
       });
 
@@ -350,30 +354,30 @@ describe('Storage', function () {
       });
     });
 
-    it('should have a filename property', function () {
+    it('should have a filename property', () => {
       expect(result.file).to.have.a.property('filename');
       expect(result.file.filename).to.be.a('string');
       expect(result.file.filename).to.match(/^[0-9a-f]{32}$/);
     });
 
-    it('should have a metadata property', function () {
+    it('should have a metadata property', () => {
       expect(result.file).to.have.a.property('metadata');
       expect(result.file.metadata).to.have.a.property('foo');
       expect(result.file.metadata.foo).to.equal('bar');
     });
 
-    it('should have a id property', function () {
+    it('should have a id property', () => {
       expect(result.file).to.have.a.property('id');
       expect(result.file.id).to.equal(1234);
     });
 
-    it('should have the default bucket name pointing to the fs collection', function () {
+    it('should have the default bucket name pointing to the fs collection', () => {
       expect(result.file).to.have.a.property('bucketName');
       expect(result.file.bucketName).to.equal('fs');
     });
 
-    it('should have the date of the upload', function () {
-      it('should have the default bucket name pointing to the fs collection', function () {
+    it('should have the date of the upload', () => {
+      it('should have the default bucket name pointing to the fs collection', () => {
         expect(result.file).to.have.a.property('uploadDate');
       });
     });
@@ -381,22 +385,22 @@ describe('Storage', function () {
     after(() => cleanStorage(storage));
   });
 
-  describe('Using empty values as return values', function () {
+  describe('Using empty values as return values', () => {
     before((done) => {
       const values = [null, undefined, {}];
       let counter = -1;
 
       storage = GridFsStorage({
-        url: setting.mongoUrl(),
+        url: setting.mongoUrl,
         file: () => {
           counter++;
           return values[counter];
-        }
+        },
       });
-      const upload = multer({ storage });
+      const upload = multer({storage});
 
       app.post('/empty', upload.array('photo', 3), (req, res) => {
-        result = { headers: req.headers, files: req.files, body: req.body };
+        result = {headers: req.headers, files: req.files, body: req.body};
         res.end();
       });
 
@@ -410,25 +414,25 @@ describe('Storage', function () {
       });
     });
 
-    it('should have the default filename', function () {
+    it('should have the default filename', () => {
       result.files.forEach((file) => {
         expect(file.filename).to.match(/^[0-9a-f]{32}$/);
       });
     });
 
-    it('should have the default metadata', function () {
+    it('should have the default metadata', () => {
       result.files.forEach((file) => {
         expect(file.metadata).to.equal(null);
       });
     });
 
-    it('should have the default bucket name', function () {
+    it('should have the default bucket name', () => {
       result.files.forEach((file) => {
         expect(file.bucketName).to.equal('fs');
       });
     });
 
-    it('should have the default chunkSize', function () {
+    it('should have the default chunkSize', () => {
       result.files.forEach((file) => {
         expect(file.chunkSize).to.equal(261120);
       });
@@ -437,22 +441,22 @@ describe('Storage', function () {
     after(() => cleanStorage(storage));
   });
 
-  describe('Using strings or numbers as return values', function () {
+  describe('Using strings or numbers as return values', () => {
     before((done) => {
       const values = ['name', 10];
       let counter = -1;
 
       storage = GridFsStorage({
-        url: setting.mongoUrl(),
+        url: setting.mongoUrl,
         file: () => {
           counter++;
           return values[counter];
-        }
+        },
       });
-      const upload = multer({ storage });
+      const upload = multer({storage});
 
       app.post('/values', upload.array('photo', 2), (req, res) => {
-        result = { headers: req.headers, files: req.files, body: req.body };
+        result = {headers: req.headers, files: req.files, body: req.body};
         res.end();
       });
 
@@ -465,24 +469,24 @@ describe('Storage', function () {
       });
     });
 
-    it('should have given filename', function () {
+    it('should have given filename', () => {
       expect(result.files[0].filename).to.equal('name');
       expect(result.files[1].filename).to.equal('10');
     });
 
-    it('should have the default metadata', function () {
+    it('should have the default metadata', () => {
       result.files.forEach((file) => {
         expect(file.metadata).to.equal(null);
       });
     });
 
-    it('should have the default bucket name', function () {
+    it('should have the default bucket name', () => {
       result.files.forEach((file) => {
         expect(file.bucketName).to.equal('fs');
       });
     });
 
-    it('should have the default chunkSize', function () {
+    it('should have the default chunkSize', () => {
       result.files.forEach((file) => {
         expect(file.chunkSize).to.equal(261120);
       });